improve queries code structure
